some cleanup and ordering (#8660)
