+++ tgui/packages/tgui/interfaces/chompstation/VorePanel/VoreSelectedBellyDescriptionTexts/VoreSelectedBellyDescriptionsStruggle.tsx
@@ -1,0 +1,39 @@
+import { useBackend } from '../../../../backend';
+import { Button, LabeledList } from '../../../../components';
+
+export const VoreSelectedBellyDescriptionsStruggle = (props) => {
+  const { act } = useBackend();
+
+  return (
+    <LabeledList.Item label="Struggle Messages">
+      <Button
+        onClick={() =>
+          act('set_attribute', { attribute: 'b_msgs', msgtype: 'smo' })
+        }
+      >
+        Struggle Message (outside)
+      </Button>
+      <Button
+        onClick={() =>
+          act('set_attribute', { attribute: 'b_msgs', msgtype: 'smi' })
+        }
+      >
+        Struggle Message (inside)
+      </Button>
+      <Button
+        onClick={() =>
+          act('set_attribute', { attribute: 'b_msgs', msgtype: 'asmo' })
+        }
+      >
+        Absorbed Struggle Message (outside)
+      </Button>
+      <Button
+        onClick={() =>
+          act('set_attribute', { attribute: 'b_msgs', msgtype: 'asmi' })
+        }
+      >
+        Absorbed Struggle Message (inside)
+      </Button>
+    </LabeledList.Item>
+  );
+};
