+++ tgui/packages/tgui/interfaces/chompstation/VorePanel/VoreSelectedBellyTabs/VoreSelectedBellyControls.tsx
@@ -1,7 +1,7 @@
-import { useBackend } from '../../../backend';
-import { Button, LabeledList } from '../../../components';
-import { digestModeToColor } from './constants';
-import { selectedData } from './types';
+import { useBackend } from '../../../../backend';
+import { Button, LabeledList } from '../../../../components';
+import { digestModeToColor } from '../constants';
+import { selectedData } from '../types';
 
 export const VoreSelectedBellyControls = (props: { belly: selectedData }) => {
   const { act } = useBackend();
