+++ tgui/packages/tgui/interfaces/chompstation/VorePanel/VoreSelectedBellyTabs/VoreSelectedBellyLiquidMessages.tsx
@@ -1,6 +1,6 @@
-import { useBackend } from '../../../backend';
-import { Button, LabeledList, Section } from '../../../components';
-import { selectedData } from './types';
+import { useBackend } from '../../../../backend';
+import { Button, LabeledList, Section } from '../../../../components';
+import { selectedData } from '../types';
 
 export const VoreSelectedBellyLiquidMessages = (props: {
   belly: selectedData;
