+++ tgui/packages/tgui/interfaces/chompstation/VorePanel/VoreSelectedBellyTabs/VoreSelectedBellyLiquidOptions.tsx
@@ -1,8 +1,8 @@
-import { useBackend } from '../../../backend';
-import { Button, LabeledList, Section } from '../../../components';
-import { reagentToColor } from './constants';
-import { LiquidColorInput } from './LiquidColorInput';
-import { selectedData } from './types';
+import { useBackend } from '../../../../backend';
+import { Button, LabeledList, Section } from '../../../../components';
+import { reagentToColor } from '../constants';
+import { LiquidColorInput } from '../LiquidColorInput';
+import { selectedData } from '../types';
 
 export const VoreSelectedBellyLiquidOptions = (props: {
   belly: selectedData;
