+++ tgui/packages/tgui/interfaces/chompstation/VorePanel/VoreSelectedBellyTabs/VoreSelectedBellyVisuals.tsx
@@ -1,9 +1,15 @@
 import { classes } from 'common/react';
 
-import { useBackend } from '../../../backend';
-import { Box, Button, Flex, LabeledList, Section } from '../../../components';
-import { FeatureColorInput } from './FeatureColorInput';
-import { selectedData } from './types';
+import { useBackend } from '../../../../backend';
+import {
+  Box,
+  Button,
+  Flex,
+  LabeledList,
+  Section,
+} from '../../../../components';
+import { FeatureColorInput } from '../FeatureColorInput';
+import { selectedData } from '../types';
 
 export const VoreSelectedBellyVisuals = (props: { belly: selectedData }) => {
   const { act } = useBackend();
