+++ tgui/packages/tgui/interfaces/chompstation/VorePanel/types.ts
@@ -5,7 +5,7 @@ export type Data = {
   show_pictures: BooleanLike;
   icon_overflow: BooleanLike;
   inside: insideData;
-  host_mobtype: { is_cyborg: BooleanLike; is_vore_simple_mob: BooleanLike };
+  host_mobtype: hostMob;
   our_bellies: bellyData[];
   selected: selectedData;
   prefs: prefData;
@@ -17,6 +17,11 @@ export type Data = {
     maximum_size: number;
     resize_cost: number;
   };
+};
+
+export type hostMob = {
+  is_cyborg: BooleanLike;
+  is_vore_simple_mob: BooleanLike;
 };
 
 export type insideData = {
@@ -158,7 +163,7 @@ export type selectedData = {
   liq_messages: liqMessageData;
 };
 
-type interactData = {
+export type interactData = {
   escapechance: number;
   escapechance_absorbed: number;
   escapetime: number;
@@ -170,7 +175,7 @@ type interactData = {
   digestchance: number;
 };
 
-type autotransferData = {
+export type autotransferData = {
   autotransferchance: number;
   autotransferwait: number;
   autotransferlocation: string;
